1. Added Logo Gauge 2. fixed internal pulldown for snooze 3. Read speed and RPM to int

+++ src/parseCan.h
@@ -6,10 +6,10 @@ void parseCanMessage(uint32_t id, const uint8_t message[], uint8_t messageLength
 int checkNumClicks(void);
 
 extern volatile float voltage;
-extern volatile float speed;
+extern volatile int speed;
 extern volatile float torquenm;
 extern volatile float torquelbf;
-extern volatile float rpm; 
+extern volatile int rpm; 
 extern volatile float coolantC;
 extern volatile float coolantF;
 extern volatile int dimmer;
